requirements.txt
Django>=3.0.1,<4

beautifulsoup4
django-bootstrap-form>=3.3,<4
django_ckeditor
django-countries
django-markdown-deux
email-reply-parser
lxml
python-magic
pytz
shortuuid
simplejson
six

# datatables ajax
git+https://github.com/peppelinux/django-datatables-ajax.git

# development
ipython

#Django form builder
#git+https://github.com/UniversitaDellaCalabria/django-form-builder.git
django-form-builder>=0.8.0

# Javascript Django Formset
# django-formset-js-improved

# FileSignatureValidator
git+https://github.com/peppelinux/FileSignatureValidator.git

# nested admin inlines
django-nested-admin

# SAML2
git+https://github.com/peppelinux/pysaml2.git@pplnx-v5
git+https://github.com/peppelinux/djangosaml2.git@pplnx-dev

# SASS dynamic compiler
libsass
git+https://github.com/jrief/django-sass-processor.git

# settings dependencies
mysqlclient

#Bootstrap Italia + Unical template
#libsass
#django-sass-processor
#django_compressor
#git+https://github.com/italia/design-django-theme.git
#git+https://github.com/UniversitaDellaCalabria/django-unical-agid-template.git
design-django-theme
django-unical-bootstrap-italia

# Django REST
djangorestframework
markdown
django-filter

# Django channels
channels
channels_redis

